refactor and package
