+++ app/src/main/AndroidManifest.xml
@@ -16,7 +16,7 @@
         android:usesCleartextTraffic="true"
         tools:targetApi="31">
         <activity
-            android:name=".WholeImageActivity"
+            android:name=".wholeimage.WholeImageActivity"
             android:exported="false" />
         <activity
             android:name=".dogdetail.DogDetailActivity"
