+++ app/src/main/java/com/navi/dogbreedapp/dogdetail/DogDetailActivity.kt
@@ -3,7 +3,7 @@ package com.navi.dogbreedapp.dogdetail
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import coil.load
-import com.navi.dogbreedapp.DogModel
+import com.navi.dogbreedapp.model.DogModel
 import com.navi.dogbreedapp.R
 import com.navi.dogbreedapp.databinding.ActivityDogDetailBinding
