+++ app/src/main/java/com/navi/dogbreedapp/doglist/DogListViewModel.kt
@@ -4,7 +4,7 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.navi.dogbreedapp.DogModel
+import com.navi.dogbreedapp.model.DogModel
 import com.navi.dogbreedapp.api.responses.ApiResponseStatus
 import kotlinx.coroutines.launch
