+++ app/src/main/java/com/navi/dogbreedapp/wholeimage/WholeImageActivity.kt
@@ -1,4 +1,4 @@
-package com.navi.dogbreedapp
+package com.navi.dogbreedapp.wholeimage
 
 import android.net.Uri
 import androidx.appcompat.app.AppCompatActivity
