+++ app/src/main/res/layout/activity_dog_detail.xml
@@ -6,7 +6,7 @@
     <data>
         <variable
             name="dog"
-            type="com.navi.dogbreedapp.DogModel" />
+            type="com.navi.dogbreedapp.model.DogModel" />
     </data>
