+++ app/src/main/res/layout/activity_whole_image.xml
@@ -9,7 +9,7 @@
     <FrameLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".WholeImageActivity">
+        tools:context=".wholeimage.WholeImageActivity">
 
         <ImageView
             android:id="@+id/whole_image_view"
